# product_manager_api_doc.yml
...
swagger: '2.0'
info:
  version: "1.0.0"
  title: Product Manager - Gestor de Categorias y Productos en una Tienda
  description: |
    </br>
    API para Gestionar Categorias y Productos.
    </br>
    </br>
  contact: 
      name: API Support
      url: https://github.com/jorgeMorfinezM/product_catalog_api/
      email: jorge.morfinez.m@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
host:
  localhost:8080
basePath: "/api/v1.0/manager"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /product:
    post:
      tags:
        - "Manage Products"
      description:
        Endpoint to insert Products (manage).
      parameters:
        - name: RequestProduct
          in: body
          description: Payload with Product data to insert.
          required: true
          schema:
            $ref: '#/definitions/RequestProduct'
      responses:
        200:
          description: Successful response Product Register
          schema:
            title: ResponseProducts
            type: array
            items:
              $ref: '#/definitions/ResponseProducts'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
    put:
      tags:
        - "Manage Products"
      description:
        Endpoint to manage data for Products (Update).
      parameters:
        - name: RequestProduct
          in: body
          description: Payload with data to updated.
          required: true
          schema:
            $ref: '#/definitions/RequestProduct'
      responses:
        200:
          description: Successful response Product updated
          schema:
            title: ResponseProducts
            type: array
            items:
              $ref: '#/definitions/ResponseProducts'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
    delete:
      tags:
        - "Manage Products"
      description:
        Endpoint to manage data for Categories of Products (Delete).
      parameters:
        - name: RequestProduct
          in: body
          description: Payload with data to inactivate.
          required: true
          schema:
            $ref: '#/definitions/RequestProduct'
      responses:
        200:
          description: Successful response Inactivate Product
          schema:
            title: ResponseProducts
            type: array
            items:
              $ref: '#/definitions/ResponseProducts'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
    get:
      tags:
        - "Manage Products"
      description:
        Get data Products.
      parameters:
        - name: Pagination
          in: body
          description: Payload with pagination data to Get.
          required: false
          schema:
            $ref: '#/definitions/Pagination'
      responses:
        200:
          description: Successful response Product Register
          schema:
            title: ResponseProducts
            type: array
            items:
              $ref: '#/definitions/ResponseProducts'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
  /product/filter/:
    get:
      tags:
        - "Retrieve Products by Filters"
      description:
        Get All Products registered by filters.
      parameters:
        - name: FilterProduct
          in: body
          description: Payload with filters to find Towns.
          required: true
          schema:
            $ref: '#/definitions/FilterProduct'
      responses:
        200:
          description: Successful response
          schema:
            title: ResponseProducts
            type: array
            items:
              $ref: '#/definitions/ResponseProducts'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
definitions:
  Error:
    allOf:
      - $ref: '#/definitions/Error'
      - required:
        - error_code
        - error_message
        type: "object"
        properties:
          error_code:
            type: integer
            format: int32
          error_message:
            type: string
...
  Pagination:
    allOf:
      - $ref: '#/definitions/Pagination'
      - required:
        - offset
        - limit
        type: "object"
        properties:
          offset:
            type: integer
            format: int64
          limit:
            type: integer
            format: int64
...
  RequestProduct:
    allOf:
      - $ref: '#/definitions/RequestProduct'
      - required:
          - sku_producto
          - marca_producto
          - unidad_medida_producto
          - inventario_producto
          - nombre_producto
          - titulo_producto
          - descripcion_larga
          - url_imagen
          - precio_unitario
          - costo_impuesto
          - tasa_impuesto
          - estatus_producto
          - producto_publicado
          - volumetria_largo_producto
          - volumetria_ancho_producto
          - volumetria_alto_producto
          - volumetria_peso_producto
          - id_categoria_producto
          - id_categoria_padre_producto
        type: "object"
        properties:
          sku_producto:
            type: string
          marca_producto:
            type: string
          unidad_medida_producto:
            type: string
          inventario_producto:
            type: integer
            format: int64
          nombre_producto:
            type: string
          titulo_producto:
            type: string
          descripcion_larga:
            type: string
          url_imagen:
            type: string
          precio_unitario:
            type: integer
            format: int64
          costo_impuesto:
            type: integer
            format: int64
          tasa_impuesto:
            type: integer
            format: int64
          estatus_producto:
            type: string
          producto_publicado:
            type: string
          volumetria_largo_producto:
            type: integer
            format: int64
          volumetria_ancho_producto:
            type: integer
            format: int64
          volumetria_alto_producto:
            type: integer
            format: int64
          volumetria_peso_producto:
            type: integer
            format: int64
          id_categoria_producto:
            type: integer
            format: int64
          id_categoria_padre_producto:
            type: integer
            format: int64
  ResponseProducts:
    allOf:
      - $ref: '#/definitions/ResponseProducts'
      - required:
          - id_producto
          - sku_producto
          - marca_producto
          - unidad_medida_producto
          - inventario_producto
          - nombre_producto
          - titulo_producto
          - descripcion_larga
          - url_imagen
          - precio_unitario
          - costo_impuesto
          - tasa_impuesto
          - estatus_producto
          - producto_publicado
          - volumetria_largo_producto
          - volumetria_ancho_producto
          - volumetria_alto_producto
          - volumetria_peso_producto
          - id_categoria_producto
          - id_categoria_padre_producto
          - cambio_realizado
          - contador_busqueda
          - fecha_alta
          - fecha_actualizacion
        type: "object"
        properties:
          id_producto:
            type: integer
            format: int64
          sku_producto:
            type: string
          marca_producto:
            type: string
          unidad_medida_producto:
            type: string
          inventario_producto:
            type: string
          nombre_producto:
            type: string
          titulo_producto:
            type: string
          descripcion_larga:
            type: string
          url_imagen:
            type: string
          precio_unitario:
            type: string
          costo_impuesto:
            type: string
          tasa_impuesto:
            type: string
          estatus_producto:
            type: string
          producto_publicado:
            type: string
          volumetria_largo_producto:
            type: string
          volumetria_ancho_producto:
            type: string
          volumetria_alto_producto:
            type: string
          volumetria_peso_producto:
            type: string
          id_categoria_producto:
            type: string
          id_categoria_padre_producto:
            type: string
          cambio_realizado:
            type: string
          contador_busqueda:
            type: string
          fecha_alta:
            type: string
          fecha_actualizacion:
            type: string
  FilterProduct:
    allOf:
      - $ref: '#/definitions/FilterProduct'
      - required:
          - sku_producto
          - marca_producto
          - nombre_producto
          - titulo_producto
          - estatus_producto
        type: "object"
        properties:
          sku_producto:
            type: string
          marca_producto:
            type: string
          nombre_producto:
            type: string
          titulo_producto:
            type: string
          estatus_producto:
            type: string